Simplify the setup process for local installs to use a prefixed miniconda install
Before this, we were checking to see if conda was already installed and was the
correct version before asking people to install the correct version of conda.
However, if conda was not installed at all, we got errors while running
`check_for_conda`.

```
check_for_conda.sh: line 2: conda: command not found
```

And while activating, we assumed that conda was installed
with the correct prefix anyway.

So let's just use the prefix based method for everything for now, which
simplifies the codebase.
- We no longer need check_for_conda
- We can just check for the file existence in activate_conda
- We can use activate_conda in both setup_notebook and activate_notebook

TODO: Unify with the e-mission server code and the public dashboard code and
create a standard module that all of them can use.